11/14/2015 Update
1. Background added: using Stomach Game BackGround Temporary.
2. Poop Control adjusted: More like Flappy Bird, which gives the player
a constant velocity every time we tap, instead of a constant force(
ignoring the acceleration)
3. Wall Added: Top and bottom Walls Added. Player will be stopped by two
invisible Wall(We will just need to add the art of the wall on the
background than)
4. Water Subtraction Zone added: Subtract the water very time the poo
collided with it. Can not move for now.

What to do next: add Bacteria, which should fill the water.